deleting a transaction rolls it back

+++ examples/buy.py
@@ -1,6 +1,7 @@
 import ConfigParser
 import sys
 import nupay
+from decimal import Decimal
 
 config = ConfigParser.RawConfigParser()
 config.read(sys.argv[1])
@@ -10,7 +11,7 @@ tokens = nupay.read_tokens_from_file(sys.argv[2])
 with nupay.SessionManager(config).create_session() as session:
     session.validate_tokens(tokens)
     try:
-        session.cash(float(sys.argv[3]))
+        session.cash(Decimal(sys.argv[3]))
     except nupay.NotEnoughCreditError as e:
         print("You need %.02f Eur extra credit for this."%e[0][1])
     print("Your total is %.02f Eur"%session.total)
